file names changed, README updated
